adding changes requested
