+++ src/app/Analytics/page.tsx
@@ -1,0 +1,9 @@
+import Analytics from "@/pages/SystemManagement/Analytics"
+
+export default function AnalyticsPage() {
+    return (
+        <>
+            <Analytics />
+        </>
+    )
+}
